Framebuffer, GUI, WAD reading changes
The most obvious change in this version is the addition of the HUD
overlay. I also draw a pause menu if you pause the game (however you
can't actually click any of the buttons!). I also switched to using a
custom framebuffer to render the scene. I also changed the WAD reading
a little bit, so that all the data is read in at the start.